add callback-based api on client core and make some redesign

+++ utils/d_string.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string.h>
+#include <stdint.h>
 
 typedef enum {
 	S_OK,
@@ -33,7 +34,7 @@ static inline void ds_free(String *s) {
 }
 
 // Internal function to grow the string's capacity, returns ERR_NOMEM on failure
-static inline short ds_grow(String *s, size_t new_cap) {
+static inline int8_t ds_grow(String *s, size_t new_cap) {
 	void *tmp = realloc(s->data, new_cap);
 	if (!tmp) return S_ERR_NOMEM;
 
@@ -42,7 +43,7 @@ static inline short ds_grow(String *s, size_t new_cap) {
 	return S_OK;
 }
 
-static inline short ds_append_char(String *s, char c) {
+static inline int8_t ds_append_char(String *s, char c) {
 	if (s->len + 2 > s->cap) {
 		size_t new_cap = s->cap ? s->cap * 2 : 16;
 		if (ds_grow(s, new_cap)) return S_ERR_NOMEM;
@@ -52,7 +53,7 @@ static inline short ds_append_char(String *s, char c) {
 	return S_OK;
 }
 
-static inline short ds_append_str(String *s, const char *str) {
+static inline int8_t ds_append_str(String *s, const char *str) {
 	size_t add = strlen(str);
 	if (s->len + add + 1 > s->cap)
 		if (ds_grow(s, (s->len + add + 1) * 2)) return S_ERR_NOMEM;
